Refactor the creation of ResolvedVariant up to DefaultArtifactSelector
