+++ subprojects/dependency-management/src/main/java/org/gradle/internal/resolve/resolver/DefaultArtifactSelector.java
@@ -16,25 +16,32 @@
 
 package org.gradle.internal.resolve.resolver;
 
+import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Maps;
 import org.gradle.api.artifacts.component.ComponentArtifactIdentifier;
+import org.gradle.api.artifacts.component.ComponentIdentifier;
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.artifact.ArtifactSet;
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.artifact.ArtifactSetFactory;
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.artifact.FileDependencyArtifactSet;
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.artifact.ResolvableArtifact;
+import org.gradle.api.internal.artifacts.ivyservice.resolveengine.artifact.ResolvedVariant;
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.simple.DefaultExcludeFactory;
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.specs.ExcludeSpec;
 import org.gradle.api.internal.artifacts.type.ArtifactTypeRegistry;
+import org.gradle.api.internal.attributes.AttributesSchemaInternal;
 import org.gradle.api.internal.attributes.ImmutableAttributes;
 import org.gradle.internal.component.local.model.LocalFileDependencyMetadata;
 import org.gradle.internal.component.model.ComponentArtifactMetadata;
 import org.gradle.internal.component.model.ComponentResolveMetadata;
 import org.gradle.internal.component.model.ConfigurationMetadata;
+import org.gradle.internal.component.model.ModuleSources;
+import org.gradle.internal.component.model.VariantResolveMetadata;
 import org.gradle.internal.model.CalculatedValueContainerFactory;
 
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 public class DefaultArtifactSelector implements ArtifactSelector {
     private static final ExcludeSpec EXCLUDE_NONE = new DefaultExcludeFactory().nothing();
@@ -61,7 +68,7 @@ public class DefaultArtifactSelector implements ArtifactSelector {
     public ArtifactSet resolveArtifacts(ComponentResolveMetadata component, ConfigurationMetadata configurationMetadata, ExcludeSpec exclusions, ImmutableAttributes overriddenAttributes) {
         ArtifactSet artifacts = null;
         for (OriginArtifactSelector selector : selectors) {
-            artifacts = selector.resolveArtifacts(new OriginArtifactSelector.DefaultComponentResolveMetadataForArtifactSelection(component, configurationMetadata.getVariants()), artifactTypeRegistry, exclusions, overriddenAttributes);
+            artifacts = selector.resolveArtifacts(new DefaultComponentResolveMetadataForArtifactSelection(component, exclusions, configurationMetadata.getVariants()), artifactTypeRegistry, exclusions, overriddenAttributes);
             if (artifacts != null) {
                 break;
             }
@@ -75,5 +82,42 @@ public class DefaultArtifactSelector implements ArtifactSelector {
     @Override
     public ArtifactSet resolveArtifacts(ComponentResolveMetadata component, Collection<? extends ComponentArtifactMetadata> artifacts, ImmutableAttributes overriddenAttributes) {
         return ArtifactSetFactory.adHocVariant(component.getId(), component.getModuleVersionId(), artifacts, component.getSources(), EXCLUDE_NONE, component.getAttributesSchema(), artifactResolver, allResolvedArtifacts, artifactTypeRegistry, component.getAttributes(), overriddenAttributes, calculatedValueContainerFactory);
+    }
+
+    class DefaultComponentResolveMetadataForArtifactSelection implements OriginArtifactSelector.ComponentResolveMetadataForArtifactSelection {
+        private final ComponentResolveMetadata delegate;
+        private final Set<? extends VariantResolveMetadata> variants;
+        private final ExcludeSpec exclusions;
+
+        public DefaultComponentResolveMetadataForArtifactSelection(ComponentResolveMetadata delegate, ExcludeSpec exclusions, Set<? extends VariantResolveMetadata> variants) {
+            this.delegate = delegate;
+            this.exclusions = exclusions;
+            this.variants = variants;
+        }
+
+        @Override
+        public ComponentIdentifier getId() {
+            return delegate.getId();
+        }
+
+        @Override
+        public ModuleSources getSources() {
+            return delegate.getSources();
+        }
+
+        @Override
+        public AttributesSchemaInternal getAttributesSchema() {
+            return delegate.getAttributesSchema();
+        }
+
+        @Override
+        public Set<ResolvedVariant> getResolvedVariants() {
+            ImmutableSet.Builder<ResolvedVariant> result = ImmutableSet.builder();
+            for (VariantResolveMetadata variant : variants) {
+                ResolvedVariant resolvedVariant = ArtifactSetFactory.toResolvedVariant(variant, delegate.getModuleVersionId(), getSources(), exclusions, artifactResolver, allResolvedArtifacts, artifactTypeRegistry, calculatedValueContainerFactory);
+                result.add(resolvedVariant);
+            }
+            return result.build();
+        }
     }
 }
